Use the "tor" role for pt0 peers in t0-isolated
Use the "tor" role for pt0 peers in the following topologies:

- t0-isolated-d256u256s2
- t0-isolated-d128u128s2
- t0-isolated-d16u16s2

While developing another testbed enhancement involving the peer devices
and configuration of t0-isolated and t0-isolated-v6 topologies, it was
noticed that the topology definitions for t0-isolated-d16u16s2,
t0-isolated-d32u32s2, t0-isolated-d128u128s2, and
t0-isolated-d256u256s2, and their v6 counterparts, do not specify that
pt0 peers use configuration for the "tor" role.  This distinction proved
beneficial for the convergence of peer docker containers (more on that
to come in future PRs), and also brought the models defined with
these topologies in-line with the definition for t0-isolated-d2u510s2.
The topology generator script is also modified such that future
t0-isolated topologies will specify the "tor" role for pt0 peers in the
future.

By request, this change is not applied to the t0-isolated-d32u32s2 or
t0-isolated-v6-d32u32s2 topologies in this branch.