Send more verification packets in packet trimming tests
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x]  202412
- [x] 202505

### Approach
#### What is the motivation for this PR?
Verification for packet trimming tests was failing on th5. We found that trimming was not occurring because too many packets were leaking out of the egress queue despite the shaper being applied.

#### How did you do it?
By increasing the number of packets sent from 10 to 100 in the verification stage, we saw that trimming was occurring and the verification passed.

#### How did you verify/test it?
Verified that packet trimming tests were passing with the change.

#### Any platform specific information?
This issue was discovered when trying to support broadcom th5.

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->