[SmartSwitch] Restore dash smartswitch vnet after protobuf updates and pr 19700 merge
### Description of PR
This change restores the functionality of the dash smartswitch vnet testing after the latest changes to the proto buffers and pr 19700
The issue was that the vnet smartswitch test was failing becase the dash parser is now using the same method as for private link but is missing some cases used when parsing config templates for vnet.
Also some of the configurations from templates need to be updated based on the latest dash proto updates.

Summary:
Fixes # (issue)

### Type of change

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505
- [x] 202506
### Approach
#### What is the motivation for this PR?

#### How did you do it?
Updated the test file and the proto file to handle the vnet configuration templates.

#### How did you verify/test it?
Executed the dash smartswitch vnet test
------------------------------------------------------------------------------------ generated xml file: /run_logs/sample/dash/test_dash_smartswitch_vnet.xml ------------------------------------------------------------------------------------
INFO:root:Can not get Allure report URL. Please check logs
=================================================================================================== 4 passed, 17 warnings in 450.40s (0:07:30)
#### Any platform specific information?
Smartswitch

#### Supported testbed topology if it's a new test case?

### Documentation
N/A