cEOSLab multi-vrf support: Topology script
Implement a script to be called from TestbedProcessing which converts a
testbed topology such that it may be run on a minimal set of cEOSLab
peers.  This involves the creation of new configuration metadata to be
read by ansible tasks-- the majority of BGP peering will exist in the
context of VRFs on a smaller set of host cEOSLab devices.