override config tests (masic): ignore certain tables in empty test
(This change was already implemented for the non-msaic version
of the test case.)

In the test case test_load_minigraph_with_golden_config, the function
load_minigraph_with_golden_empty_input asserts. This function expects all the
tables to be identical after the golden config is overridden to be empty.  But
it looks as if the golden config is the source of several configurations: the
"BMP" table, and "fec" : "rs" entries in the "PORT".  Thus, when the golden
config is replaced with an empty configuration, the entire "BMP" table,
well as those "fec" properties in the "PORT" table, disappear.

We propose a variable GOLDEN_OVERRIDDEN_TABLES for additional tables
which that function should skip. This is similar to the existing variable
NON_USER_CONFIG_TABLES which non-user-config tables that need to be skipped.

tests/override_config_table/test_override_config_table_masic.py
(GOLDEN_OVERRIDDEN_TABLES): New variable.
(load_minigraph_with_golden_empty_input): Skip tables found in
GOLDEN_OVERRIDDEN_TABLES.