sonic-mgmt: Infrastructure changes for multi-vrf
Introduce infrastructure changes required to converge multiple BGP peers
into a minimum number of cEOSLab hosts, via the use of VLANs and VRFs.

Overview of peer convergence:

Converging the total number of peer switches into the fewest possible
number of cEOSLab containers reduces the overall resource constraints
required to run large numbers of peers.  The basic premises behind
convergence are as follows:

- cEOSLab peers in docker containers may be converged into a smaller
  number of host peers.
- The SONiC-facing configuration of each BGP peer may be separated in
  routing and bridging via the use of VRFs.
- The PTF-facing configuration of each BGP peer may be separated within
  each VRF via VLAN tagging, enabling the use of a single backplane
  interface on each host cEOSLab container.
- Each VRF includes a number of interfaces either facing the SONiC DUT
  or the backplane.
- Changes should be as transparent to the SONiC DUT as possible.

At the time of testbed setup, the ansible topology file for the testbed
is modified to include new metadata specific to multi-vrf configuration,
and the VMs list is trimmed to only include those containers which will
host multiple BGP peerings, separated by VRF.  The new metadata includes
mappings between host containers and VRFs, backplane VLAN mappings, and
BGP session parameters.

VLAN tag 2000 is used as the starting value for all VLANs between the
test infrastructure PTF container interfaces and cEOSLab device
interfaces.

The IP and IPv6 addresses used to connect the cEOSLab peer and
infrastructure PTF container are generated in order to
make the backplane connections clearer, more unique, and easier to
implement.  In general, backplane L3 addresses used by the CEOSLab peer
end in even numbers, and those used by the PTF container end in odd
numbers.  All addresses generated for use in backplane connections start
with the value 100 (0x64) in the least-significant octet or hextet
(depending on the family of the address).  The address changes are
mapped and stored in the new multi-vrf metadata in the ansible topology
file.

Multiple BGP features, such as local-as and next-hop-peer, are used in
order to aid in the resolution of routes.  This is necessary to keep the
SONiC DUT multi-vrf-agnostic as possible.

Known limitations:
- cEOSLab instances do not allow for the creation of interfaces with
  interface-IDs greater than 127, when interfaces are layed out
  unidimensionally.
- The use of multiple VRFs has not been tested in conjunction with
  asynchronous ansible tasks.