sonic-mgmt: Test lib changes for multi-vrf testbeds
Introduce code to test libraries specifically for running tests on
testbeds with multi-vrf cEOSLab peer devices.

These changes predominantly make test libraries aware of the host-vrf
mappings which are required for peer multi-vrf convergence.  In such a
testbed, a small number of host cEOSLab containers include configuration
for a numbe of VRFs, inside which BGP adjacencies to the PTF and SONiC
DUTs are formed.  Knowledge of these mappings is required in order to
correctly manipulate the host of a given VRF/BGP peer, as well as to
correctly interpret the SONiC LLDP tables (only host cEOSLab containers
are show via LLDP with multi-vrf enabled, which multiple connections).