Support running acl tests without any IPv4 management configuration
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fixes #18077

Modify the gratuitous arp service such that arp or neighbor-discovery packets are sent based on the L3 configuration of the testbed topology.  For testbed topologies which do not configure IPv4 addressing, such as isolated-v6 testbeds, we skip the creation of arp packets and do not send them.  In this case, only IPv6 neighbor-discovery packets are created and sent.  The inverse is true as well when a testbed topology only specifies the use of IPv4 connections.

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?
This change was made as a part of the isolated-v6 testbed qualification effort.  Running acl tests on isolated-v6 testbeds was previously unsupported, and skipped via conditional mark, as the tests expected to receive IPv4 arp packets when no IPv4 connectivity was established.

#### How did you verify/test it?
The full suite of acl tests was run against a testbed where IPv4 and IPv6 connectivity was specified in the topology config, and no regressions were seen.  The same suite of tests was run against a isolated-v6 testbed, and new test passes were observed: some new failures were also observed for tests which were previously skipped.  However @r12f asked that we submit this change, and follow up with the remaining failing tests individually in subsequent changes as this change improves overall test coverage.

#### Any platform specific information?
None.